[CDEC-345] Refactor Pos.Core.Delegation module set
